fix(linux): napraw błędy kompilacji GCC 15 / Qt 6.10 na Linux
- SocketCanDriver: std::min(uint8_t, int) → rzutowanie na int (GCC 15 ścisłe typy)
- CanObservationDb: int64_t → qlonglong przy addBindValue (na Linux int64_t=long, Qt nie ma konstruktora QVariant(long))
- CMakeLists: pkg_check_modules(ZSTD) przeniesione poza blok WIN32 — libzstd nie linkował się na Linux

+++ src/core/CanObservationDb.cpp
@@ -85,7 +85,7 @@ int64_t CanObservationDb::beginSession(const QString &label) {
     QSqlQuery q(m_db);
     q.prepare("INSERT INTO sessions (label, started) VALUES (?, ?)");
     q.addBindValue(label.isEmpty() ? QVariant{} : QVariant(label));
-    q.addBindValue(now);
+    q.addBindValue(static_cast<qlonglong>(now));
     q.exec();
     m_sessionId = q.lastInsertId().toLongLong();
     return m_sessionId;
@@ -96,8 +96,8 @@ void CanObservationDb::endSession() {
     int64_t now = QDateTime::currentMSecsSinceEpoch();
     QSqlQuery q(m_db);
     q.prepare("UPDATE sessions SET ended=? WHERE id=?");
-    q.addBindValue(now);
-    q.addBindValue(m_sessionId);
+    q.addBindValue(static_cast<qlonglong>(now));
+    q.addBindValue(static_cast<qlonglong>(m_sessionId));
     q.exec();
     m_sessionId = -1;
 }
